Adding feature:upload images file to storage, and fixing some image setting every view page

+++ app/Http/Controllers/GameController.php
@@ -3,6 +3,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 use App\Models\Category;
 use App\Models\Game;
@@ -13,6 +14,10 @@ class GameController extends Controller
     //view
     public function gameDetails($slug){
         $game = Game::where('slug', $slug)->first();
+
+        if(!$game){
+            abort(404);
+        }
 
         $morelikethis = Game::where('category_id', $game->category_id)
             ->where('id', '!=', $game->id)
@@ -32,12 +37,16 @@ class GameController extends Controller
             'recommended' => $recommended_count,
             'not' => $reviews->count()-$recommended_count,
         ];
-        // return dd($counter);
+
+        $image_paths = Storage::files('assets/'.$game->image_path);
+
         return view('game')
             ->with('vargame', $game)
             ->with('morelikethis', $morelikethis)
             ->with('reviews', $reviews)
-            ->with('counter', $counter);
+            ->with('counter', $counter)
+            ->with('image_paths', $image_paths)
+            ->with('total_images', count($image_paths));
     }
 
     public function search(Request $request){
@@ -48,6 +57,9 @@ class GameController extends Controller
     public function dashboard(){
         return view('dashboard');
     }
+    
+
+
 
     //manage-game by admin
     public function index(){
@@ -56,7 +68,8 @@ class GameController extends Controller
     }
 
     public function formAddGame(){
-        return view('input-addgame');
+        $categories = Category::all();
+        return view('add-game')->with("categories", $categories);
     }
 
     public function store(Request $request){
@@ -90,12 +103,33 @@ class GameController extends Controller
         $newGame->updated_at = now();
         $newGame->save();
 
-        return redirect()->back();
+        $path = 'assets/'.$newGame->image_path;
+        $file = $request->file('thumbnail');
+        $thumb_filename = 'thumb.'.$file->extension();
+
+        Storage::putFileAs(
+            $path, $file, $thumb_filename
+        );
+
+        $counter = 0;
+        foreach($request->file('slider') as $slider){
+            $counter++;
+            $slider->storeAs($path, 'slide'.$counter.'.'.$slider->extension());
+        }
+
+        return redirect()->route('manage-games')->with('add_success', 'Game <b>"'.$request->title.'"</b> has successfully been added!');
     }
 
     public function formUpdateGame($slug){
         $game = Game::with('category')->where('slug', $slug)->first();
-        return view('update-game')->with('game',$game);
+
+        if(!$game){
+            abort(404);
+        }
+
+        $categories = Category::all();
+
+        return view('update-game')->with('game',$game)->with("categories", $categories);
     }
 
     public function update(Request $request){
@@ -143,20 +177,40 @@ class GameController extends Controller
         $updateGame->description = $request->description;
         $updateGame->image_path = $request->title.'-img/';
         $updateGame->updated_at = now();
-
         $updateGame->save();
+        
+        Storage::deleteDirectory($request->oldTitle."-img/");
 
-        return redirect(route('manage-games'))->with('update_success', 'Game <b>"'.$request->oldTitle.'"</b> has successfully been updated!');;
+        $path = 'assets/'.$updateGame->image_path;
+        $file = $request->file('thumbnail');
+        $thumb_filename = 'thumb.'.$file->extension();
+
+        Storage::putFileAs(
+            $path, $file, $thumb_filename
+        );
+
+        $counter = 0;
+        foreach($request->file('slider') as $slider){
+            $counter++;
+            $slider->storeAs($path, 'slide'.$counter.'.'.$slider->extension());
+        }
+
+
+        return redirect()->route('manage-games')->with('update_success', 'Game <b>"'.$request->oldTitle.'"</b> has successfully been updated!');;
     }
 
     public function delete(Request $request){
         $deleteGame = Game::where('id', $request->id)->first();
-        $title = $deleteGame->title;
 
+        if(!$deleteGame){
+            abort(404);
+        }
+
+        $title = $deleteGame->title;
         $deleteGame->delete();
 
         return redirect()
-            ->back()
+            ->route('manage-games')
             ->with('delete_success', 'Game <b>"'.$title.'"</b> has successfully been deleted!');
     }
 }
